requirements.txt
streamlit==1.26.0
numpy==1.23.5
opencv-python==4.7.0.72
tensorflow==2.12.0
Pillow==9.5.0